LayoutTests/platform/glib/imported/w3c/web-platform-tests/cors/client-hint-request-headers-expected.txt

CONSOLE MESSAGE: Request header field y-print is not allowed by Access-Control-Allow-Headers.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors/resources/cors-makeheader.py?headers=x-print due to access control checks.
CONSOLE MESSAGE: Request header field device-memory is not allowed by Access-Control-Allow-Headers.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors/resources/cors-makeheader.py?headers=x-print due to access control checks.
CONSOLE MESSAGE: Request header field dpr is not allowed by Access-Control-Allow-Headers.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors/resources/cors-makeheader.py?headers=x-print due to access control checks.
CONSOLE MESSAGE: Request header field width is not allowed by Access-Control-Allow-Headers.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors/resources/cors-makeheader.py?headers=x-print due to access control checks.
CONSOLE MESSAGE: Request header field viewport-width is not allowed by Access-Control-Allow-Headers.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors/resources/cors-makeheader.py?headers=x-print due to access control checks.
Request headers


PASS Unspecified request headers are disallowed
PASS Unextractable device-memory client hint header is disallowed
PASS Unextractable DPR client hint header is disallowed
PASS Unextractable width client hint header is disallowed
PASS Unextractable viewport-width client hint header is disallowed
